Implementing new & more functional DSL, using Machinery module
The previous DSL used macros mores heavily what is known to be a bad
smell. This new DSL relies more on the Machinery module what will result into
a more clear and better documentation and a more usual implementation.
This new DSL still relies on macros but way less than the previous one.