Update Machinery DSL, decoupling it from the Struct itself.
By forcing people to send the module where the machinery was imported, as an
argument of `transition_to/2`, we make sure we are enforcing an architecture
where the State Machine logic doesn't needs to be on the Struct module itself
(what in most Web framewokrs end up being the model).
This helps ppl to avoind having a bunch o logic in their models, an old issue
known as Fat Models, while also enbaling ppl to have multiple and reusable
State Machines for different Structs.